cambio de estructura en capas por una modular mejor orden en el codigo
