+++ backend/app/models/Role/Rol_Routers.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter
-from app.schemas.Rol_Schemas import RoleBase
+from backend.app.models.Role.Rol_Schemas import RoleBase
 
 router = APIRouter()
 
@@ -9,6 +9,6 @@ def root():
     return {"msg": "mi api del rol"}
 
 
-@router.post("/")
+@router.post("/api/rol")
 def create_role(data_role: RoleBase):
     return {"msg": "Rol creado", "data": data_role}
