--- backend/app/models/__init__.py
@@ -1,1 +1,0 @@
-from .category_User import User, Role
