--- backend/app/routers/__init__.py
@@ -1,1 +1,0 @@
-from .Rol_Routers import router as Rol_Router
