--- backend/app/schemas/__init__.py
@@ -1,1 +1,0 @@
-from .Rol_Schemas import RoleBase, RoleCreate, RoleResponse
