+++ backend/tests/tests_db.py
@@ -1,4 +1,4 @@
-from backend.app.core.database import engine
+from backend.app.db.database import engine
 from sqlalchemy import text
 
 try:
